Add `backendResourceName` for model and store
When using `model.toBackendAll()`, the name of the relation is sometimes
not the same as the name of the model or store. e.g.

```js
relations() {
	return { activities: ActivityStore };
}
```

But the backend knows the `activities` relation by the name `activity`.
To fix this, you can now set `backendResourceName` on a model or store:

```js
class ActivityStore extends Store {
	static backendResourceName = 'activity';
}
```