Add support for parsing reverse relations without id list in main data
In a Binder backend, this allows us to use "with" plus a "where"
filter without returning a (potentially huge) list of IDs.  This saves
time fetching unnecessary IDs from the database, serializing them and
deserializing them.  The only requirement is that you have to use
"with" when you remove the field from m2m_fields, but that's a small
price to pay if you consider the performance benefits it brings.