Refactor toBackendAll to be faster and simpler
This represents an internal API change: toBackendAll no longer accepts
newId; it is patched up in the caller rather than the callee.

Instead of processing the relations list over and over and over,
filtering out stuff we don't need, we instead process it once in
saveAllModels before calling toBackendAll, to a nested
tree (dictionary).  This means we can simply pass the subtree to the
subrelation.

This is preparatory work to drop unnecessary serializations in Binder,
so we can prune foo: { id: 1 } relations.