Only send actually changed relations in toBackendAll
This is not 100% perfect yet: It would be better if we only serialized
the id(s) in the main object when the relation itself changed.

This is a bit trickier because that requires changes in Store, so it
tracks when ids have been added or removed.  For consistency's sake,
we always simply send the related key ID for now, even when we know
it's a related model rather than a store, and it has no changes.